+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+- Added `of(context)` extension to `ArcaneService`
+
 ## 1.2.5
 
 - Improved automatic metadata detection in `ArcaneLogger`
@@ -58,7 +62,7 @@ These methods have been moved to mixin classes. To continue using them, please
 update your ArcaneAuthInterface implementations.
 
 - To use `resendVerificationCode`, `register` and `confirmSignup`, use the new
-`ArcaneAuthAccountRegistration` mixin.
+  `ArcaneAuthAccountRegistration` mixin.
 - To use `resetPassword`, use the new `ArcaneAuthPasswordManagement` mixin.
 
 ### Migration